# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyTaskResponse:
    properties:
      board_id:
        type: string
      board_name:
        type: string
      column_name:
        type: string
      completed_subtasks:
        type: integer
      due_date:
        type: string
      estimated_hours:
        type: number
      group:
        type: string
      id:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
      total_subtasks:
        type: integer
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyWorkCounts:
    properties:
      later:
        type: integer
      no_date:
        type: integer
      overdue:
        type: integer
      this_week:
        type: integer
      today:
        type: integer
      total:
        type: integer
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyWorkResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyTaskResponse'
        type: array
      counts:
        $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyWorkCounts'
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Turtask
  description: |-
    Multi-board Kanban + task management with realtime sync.
    Auth uses an HttpOnly `auth_token` cookie issued by /api/auth/login or /api/auth/oauth.
  title: Turtask API
  version: "1.0"
paths:
...
  /api/my-tasks:
    get:
      parameters:
      - description: all|overdue|today|this_week|no_date
        in: query
        name: filter
        type: string
      - description: include unassigned cards on my boards
        in: query
        name: include_unassigned
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyWorkResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: My work (cross-board)
      tags:
      - my-tasks
  /api/my-tasks/{cardID}/complete:
    post:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: not the assignee or card missing
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Complete one of my tasks
      tags:
      - my-tasks
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"